Remove improper generics
While generics would be great in the container, the previous implementation was flawed and may be incompatible with PSR-11 as it stands today.

It's also extremely possible that I'm just misunderstanding their purpose :/

Until we can get them properly implemented, remove them and plan on `/** @type ... */` declarations in places where we're accessing Container resolutions directly.