Let make() return a cached dependency if resolutionCacheDepth > 0
This extends #11, moving the "are we currently in a caching context" check to the front of the method.

This has the benefit of making sure that we're not needlessly resolving nested dependencies, ensuring that the instance we inject into dependencies is the same as what we have in the cache.